% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_balance.R
\name{inspect_balance}
\alias{inspect_balance}
\title{Inspect class imbalance}
\usage{
inspect_balance(data, vars = NULL, fill_colour = "grey30", title = NULL, ...)
}
\arguments{
\item{data}{A date frame or tibble with at least one categorical variable.}

\item{vars}{A vector of categorical variables contained in data. Default value \code{NULL} will produce plots for every categorical variable in the data.}

\item{fill_colour}{An R-supported colour or hex code to assign as the barplot colour. Default value is "grey30"}

\item{title}{A string. Default value \code{NULL} is no title.}

\item{...}{Additional ggplot2 parameters to modify plot outputs.}
}
\value{
A ggplot2 object with bar plots for each categorical variable.
}
\description{
Quickly inspect the distribution of your categorical variables.
}
\details{
\code{inspect_balance} uses ggplot2 to produce bar plots. Any valid arguments that may be passed to a \code{geom_bar} layer may also be passed to \code{inspect_balance} to modify plot outputs.
}
\examples{
# Basic usage
inspect_balance(iris)

# Advanced usage
# Manually specifies the variable to inspect and modifies plot output
inspect_balance(data = iris,
                c("Species"),
                fill_colour = "blue",
                title = "Class balance of Species")
}
