% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_normality.R
\name{inspect_normality}
\alias{inspect_normality}
\title{Inspect normality}
\usage{
inspect_normality(data, ...)
}
\arguments{
\item{data}{A date frame or tibble with at least one numeric variable.}

\item{...}{Additional ggplot2 parameters to modify plot outputs.}
}
\value{
A faceted ggplot2 object with histograms for each numeric variable.
}
\description{
Quickly inspect the distribution of your numeric variables.
}
\examples{
# Basic usage
inspect_normality(iris)
}
