% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_missing.R
\name{inspect_missing}
\alias{inspect_missing}
\title{Inspect missingness}
\usage{
inspect_missing(
  data,
  vars = NULL,
  na_colour = "red",
  fill_colour = "darkgrey",
  ...
)
}
\arguments{
\item{data}{A non-empty date frame or tibble.}

\item{vars}{A vector of variables contained in data. Default value \code{NULL} will produce plots for every variable in the data}

\item{na_colour}{An R-supported colour or hex value to indicate missing values. Default value is "red".}

\item{fill_colour}{An R-supported colour or hex value to indicate non-missing values. Default value is "darkgrey".}

\item{...}{Additional ggplot2 parameters to modify plot outputs.}
}
\value{
A ggplot2 object.
}
\description{
Quickly inspect the missingness in your data.
}
\details{
\code{inspect_missing} uses ggplot2 to produce a bar plot. Any valid arguments that may be passed to a \code{geom_bar} layer may also be passed to \code{inspect_missing} to modify plot outputs.
}
\examples{
# Basic usage
inspect_missing(iris)

# Advanced usage
# Manually specifies the variables to inspect and modifies plot output.
inspect_missing(iris,
                c("Sepal.Width", "Sepal.Length"),
                na_colour = "transparent",
                fill_colour = "blue",
                alpha = 0.5)
}
