% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_normality.R
\name{inspect_normality}
\alias{inspect_normality}
\title{Inspect normality}
\usage{
inspect_normality(
  data,
  vars = NULL,
  fill_colour = "grey30",
  title = NULL,
  bins = 15,
  ...
)
}
\arguments{
\item{data}{A non-empty data frame or tibble containing at least one numeric variable.}

\item{vars}{A character vector of numeric variable names contained in \code{data}. Default value \code{NULL} will produce plots for every numeric variable in the dataset.}

\item{fill_colour}{An R-supported colour name or hex value used to fill the bars of the histogram. Default value is "grey30".}

\item{title}{A non-empty string for the title of the plot. Default value \code{NULL} results in no title being displayed.}

\item{bins}{A positive integer specifying the number of bins in the histogram. Default value is \code{15}.}

\item{...}{Additional \code{ggplot2} arguments passed to \code{geom_histogram} layer for customizing the plot output.}
}
\value{
A ggplot2 object containing histograms for each specified numeric variable.
}
\description{
Quickly inspect the distribution of numeric variables using histograms.
}
\details{
\code{inspect_normality} uses \code{ggplot2} to produce histograms of numeric variables to visualize their distributions. Any valid arguments that can be passed to a \code{geom_histogram} layer in ggplot2 may also be passed to \code{inspect_normality} to modify the plot outputs.
}
\examples{
# Basic usage
inspect_normality(iris)

# Advanced usage
# Specify the variables to inspect and modify the plot appearance
inspect_normality(data = iris,
                  vars = c("Sepal.Length", "Sepal.Width"),
                  fill_colour = "blue",
                  title = "Distribution of Sepal Length & Width",
                  bins = 10)
}
